- Fixed bugs with 24 bit get/set memory methods.
git-svn-id: svn://localhost/mosa@478 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/DeviceSystem/IMemory.cs
@@ -58,15 +58,14 @@ namespace Mosa.DeviceSystem
 		/// </summary>
 		/// <param name="index">The index.</param>
 		/// <param name="value">The value.</param>
-		void Write24(uint index, ushort value);
-
+		void Write24(uint index, uint value);
 
         /// <summary>
         /// Reads the specified index.
         /// </summary>
         /// <param name="index">The index.</param>
         /// <returns></returns>
-        ushort Read24(uint index);
+        uint Read24(uint index);
 
         /// <summary>
         /// Writes the specified index.
